requirements.txt
asgiref==3.5.2
backports.zoneinfo==0.2.1
certifi==2022.9.24
charset-normalizer==2.1.1
Django==4.1.3
django-filter==22.1
djangorestframework==3.14.0
idna==3.4
importlib-metadata==5.1.0
Markdown==3.4.1
pytz==2022.6
requests==2.28.1
sqlparse==0.4.3
tzdata==2022.7
urllib3==1.26.13
zipp==3.11.0
